+++ src/main/java/com/darek/customer/CustomerRepository.java
@@ -2,7 +2,7 @@ package com.darek.customer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
     boolean existsCustomerByEmail(String email);
-    boolean existsCustomerById(Integer Id);
+    boolean existsCustomerById(Long Id);
 }
